add register

+++ src/components/Navbar.jsx
@@ -82,15 +82,15 @@ function Navbar({ send, wishCount }) {
         Italy: 'https://static.vecteezy.com/system/resources/previews/011/571/348/non_2x/circle-flag-of-italy-free-png.png'
     };
 
-
     const navigate = useNavigate();
 
     const handleClearLocalStorage = () => {
-        toast.success("Log Out")
-        localStorage.clear();
-        navigate('/');
+        if (window.confirm("Are you sure you want to log out?")) {
+            toast.success("Logged Out Successfully");
+            localStorage.clear();
+            navigate('/');
+        }
     };
-
 
     return (
         <>
@@ -191,7 +191,7 @@ function Navbar({ send, wishCount }) {
                                 <svg xmlns="http://www.w3.org/2000/svg" width="20px" className="cursor-pointer fill-333 inline" viewBox="0 0 64 64">
                                     <path d="M45.5 4A18.53 18.53 0 0 0 32 9.86 18.5 18.5 0 0 0 0 22.5C0 40.92 29.71 59 31 59.71a2 2 0 0 0 2.06 0C34.29 59 64 40.92 64 22.5A18.52 18.52 0 0 0 45.5 4ZM32 55.64C26.83 52.34 4 36.92 4 22.5a14.5 14.5 0 0 1 26.36-8.33 2 2 0 0 0 3.27 0A14.5 14.5 0 0 1 60 22.5c0 14.41-22.83 29.83-28 33.14Z" data-original="#000000"></path>
                                 </svg>
-                                <span className="absolute left-auto -ml-1 top-0 rounded-full bg-red-500 px-1 py-0 text-xs text-white">{wishCount }</span>
+                                <span className="absolute left-auto -ml-1 top-0 rounded-full bg-red-500 px-1 py-0 text-xs text-white">{wishCount}</span>
                             </span>
                             <span className="relative">
                                 <Link to="/cartlist">
@@ -224,7 +224,7 @@ function Navbar({ send, wishCount }) {
                                         </li>
                                         <li className='px-4 py-2 hover:bg-rose-400 hover:text-white'>
                                             {
-                                                getEmail ? <button onClick={handleClearLocalStorage}>Logout</button> : <Link to="/contact">Sign in</Link>
+                                                getEmail ? <button onClick={handleClearLocalStorage}>Logout</button> : <Link to="/signin">Sign in/Registration</Link>
                                             }
                                         </li>
                                     </ul>
